# embedded/openapi.template.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Hubchat API",
    "version": "1.0.0",
    "description": "The versioned, workspace-scoped REST contract for Hubchat. Browser sessions use cookies; server integrations use a workspace-scoped API key as a Bearer token."
  },
  "servers": [{ "url": "/api", "description": "The Hubchat API mounted by the application" }],
...
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/v1/feedback/boards/{id}/items": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "get": { "$ref": "#/components/operations/list", "operationId": "listFeedbackItems", "tags": ["Feedback"] },
      "post": { "$ref": "#/components/operations/create", "operationId": "createFeedbackItem", "tags": ["Feedback"] }
    },
    "/v1/feedback/items/{id}/links": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "get": { "$ref": "#/components/operations/list", "operationId": "listFeedbackLinks", "tags": ["Feedback"] },
      "post": { "$ref": "#/components/operations/create", "operationId": "linkFeedbackSupportWork", "tags": ["Feedback"] }
    },
    "/v1/feedback/items/{id}/links/{linkID}": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }, { "$ref": "#/components/parameters/LinkID" }],
      "delete": { "$ref": "#/components/operations/delete", "operationId": "unlinkFeedbackSupportWork", "tags": ["Feedback"] }
    },
    "/v1/feedback/items/{id}/merge": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "post": { "$ref": "#/components/operations/create", "operationId": "mergeFeedbackItem", "tags": ["Feedback"] }
    },
    "/v1/public/feedback/{workspaceID}/boards/{slug}/items": {
      "parameters": [{ "$ref": "#/components/parameters/WorkspaceID" }, { "$ref": "#/components/parameters/Slug" }],
      "get": { "$ref": "#/components/operations/publicList", "operationId": "listPublicFeedbackItems", "security": [], "tags": ["Feedback"] },
      "post": { "$ref": "#/components/operations/publicCreate", "operationId": "createPublicFeedbackItem", "security": [], "tags": ["Feedback"] }
    },
    "/v1/public/feedback/{workspaceID}/items/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/WorkspaceID" }, { "$ref": "#/components/parameters/Id" }],
      "get": { "$ref": "#/components/operations/publicGet", "operationId": "getPublicFeedbackItem", "security": [], "tags": ["Feedback"] }
    },
    "/v1/public/feedback/{workspaceID}/items/{id}/comments": {
      "parameters": [{ "$ref": "#/components/parameters/WorkspaceID" }, { "$ref": "#/components/parameters/Id" }],
      "get": { "$ref": "#/components/operations/publicList", "operationId": "listPublicFeedbackComments", "security": [], "tags": ["Feedback"] },
      "post": { "$ref": "#/components/operations/publicCreate", "operationId": "createPublicFeedbackComment", "security": [], "tags": ["Feedback"] }
    },
    "/v1/public/feedback/{workspaceID}/items/{id}/subscription": {
      "parameters": [{ "$ref": "#/components/parameters/WorkspaceID" }, { "$ref": "#/components/parameters/Id" }],
      "post": { "$ref": "#/components/operations/publicCreate", "operationId": "subscribeToFeedbackItem", "security": [], "tags": ["Feedback"] },
      "delete": { "$ref": "#/components/operations/publicDelete", "operationId": "unsubscribeFromFeedbackItem", "security": [], "tags": ["Feedback"] }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "hc_live_...", "description": "Workspace-scoped API key. The complete token is shown only once." },
      "portalCookie": { "type": "apiKey", "in": "cookie", "name": "hubchat_portal_session", "description": "HttpOnly customer portal session cookie." }
    },
    "parameters": {
      "Id": { "name": "id", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/Id" } },
      "LinkID": { "name": "linkID", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/Id" } },
      "WorkspaceID": { "name": "workspaceID", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/Id" } },
      "Slug": { "name": "slug", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^[a-z0-9][a-z0-9-]{1,38}[a-z0-9]$" } },
      "Cursor": { "name": "cursor", "in": "query", "required": false, "schema": { "type": "string" }, "description": "Opaque cursor returned as next_cursor." },
      "Limit": { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 50 } },
      "IdempotencyKey": { "name": "Idempotency-Key", "in": "header", "required": false, "schema": { "type": "string", "minLength": 8 }, "description": "Recommended for every retryable create or action." }
    },
    "operations": {
      "list": { "summary": "List resources", "parameters": [{ "$ref": "#/components/parameters/Cursor" }, { "$ref": "#/components/parameters/Limit" }], "responses": { "200": { "$ref": "#/components/responses/Json" }, "default": { "$ref": "#/components/responses/Error" } } },
...
      "create": { "summary": "Create a resource", "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }], "requestBody": { "$ref": "#/components/requestBodies/Json" }, "responses": { "201": { "$ref": "#/components/responses/Json" }, "202": { "$ref": "#/components/responses/Json" }, "default": { "$ref": "#/components/responses/Error" } } },
...
      "delete": { "summary": "Delete a resource", "responses": { "204": { "description": "Deleted" }, "default": { "$ref": "#/components/responses/Error" } } },
      "publicGet": { "summary": "Read a published resource", "responses": { "200": { "$ref": "#/components/responses/Json" }, "default": { "$ref": "#/components/responses/Error" } } },
      "publicList": { "summary": "List published resources", "parameters": [{ "$ref": "#/components/parameters/Cursor" }, { "$ref": "#/components/parameters/Limit" }], "responses": { "200": { "$ref": "#/components/responses/Json" }, "default": { "$ref": "#/components/responses/Error" } } },
      "publicCreate": { "summary": "Create or submit a public resource", "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }], "requestBody": { "$ref": "#/components/requestBodies/Json" }, "responses": { "200": { "$ref": "#/components/responses/Json" }, "201": { "$ref": "#/components/responses/Json" }, "202": { "$ref": "#/components/responses/Json" }, "default": { "$ref": "#/components/responses/Error" } } },
      "publicDelete": { "summary": "Remove a public subscription", "responses": { "200": { "$ref": "#/components/responses/Json" }, "default": { "$ref": "#/components/responses/Error" } } },
...
    },
    "requestBodies": {
      "Json": { "required": true, "content": { "application/json": { "schema": { "type": "object", "additionalProperties": true } } } }
    },
    "responses": {
      "Json": { "description": "JSON response", "content": { "application/json": { "schema": { "type": "object", "additionalProperties": true } } } },
      "Error": { "description": "The standard Hubchat error envelope", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    },
    "schemas": {
      "Id": { "type": "string", "pattern": "^[a-z][a-z0-9]*_[0-9a-z]+$", "description": "Opaque prefixed Hubchat identifier; never parse it." },
      "Error": { "type": "object", "required": ["error"], "properties": { "error": { "$ref": "#/components/schemas/ErrorBody" } } },
      "ErrorBody": { "type": "object", "required": ["code", "message", "request_id"], "properties": { "code": { "type": "string" }, "message": { "type": "string" }, "request_id": { "$ref": "#/components/schemas/Id" }, "details": { "type": "object", "additionalProperties": true } } },
...
    }
  }
}